requirements.txt
flask==3.0.0
gunicorn==21.2.0
requests==2.31.0
yt-dlp==2023.12.30
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
